requirements.txt
flask==2.3.3
sounddevice==0.4.6
numpy==1.24.3
scipy==1.11.1
requests==2.31.0
python-dotenv==1.0.0
google-generativeai==0.8.3
